pyproject.toml
[project]
name = "research-paper-assistent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "pandas>=3.0.1",
    "pymupdf>=1.27.2",
    "pypdf>=6.9.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
